requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
slack-sdk==3.26.2
openai==1.51.2
httpx==0.27.2
python-dotenv==1.0.0